Refactor FoodCreateModal to use ngrx store
The FoodCreateModal component has been refactored to utilize ngrx store for state management, removing direct usage of the FoodsService. The process of creating a new food now dispatches an action handled by the store, improving the maintainability and scalability of the code. Also, checks for action successful completion are now handled via subscription to the selectFoodDetails selector.