Refactor FoodListBase class to include ngRx store
The FoodListBase class was refactored to include injections and subscriptions to the ngRx store and its selectAllFoods selector. The changes remove the need to inject and subscribe to the store manually in all the child classes, making the code cleaner and more maintainable. Manual injections and subscriptions, previously in FoodListPage and FoodListMobile, have been removed in this commit.